fix: split module index into client and server entry points
Single index.ts files re-exported .server modules alongside UI
components, causing "Server-only module referenced by client" build
errors. Each module now has index.ts (client-safe) and index.server.ts
(server-only). Updated all route, test, and cross-module imports.
