+++ app/modules/drinks/implementation/mutations.server.ts
@@ -1,6 +1,6 @@
 import type { z } from "zod";
 import type { Drink } from "#/app/db/schema";
-import { purgeSearchCache } from "#/app/modules/search";
+import { purgeSearchCache } from "#/app/modules/search/index.server";
 import {
   createDrink as insertDrink,
   updateDrink as patchDrink,
