+++ app/modules/drinks/index.server.ts
@@ -1,0 +1,16 @@
+// Queries
+export {
+  getAllDrinks,
+  getPublishedDrinks,
+  getDrinkBySlug,
+  getDrinksByTag,
+  getAllTags,
+} from "./implementation/queries.server";
+
+// Mutations (orchestrate DB + ImageKit + Fastly + search cache)
+export { createDrink, updateDrink, deleteDrink } from "./implementation/mutations.server";
+
+// View helpers
+export { withPlaceholderImages } from "./implementation/placeholder-images.server";
+export { parseImageUpload } from "./implementation/parse-image-upload.server";
+export { markdownToHtml } from "./implementation/markdown.server";
