+++ app/modules/drinks/index.ts
@@ -1,25 +1,8 @@
 // Types
 export type { EnhancedDrink } from "./implementation/types";
 
-// Queries
-export {
-  getAllDrinks,
-  getPublishedDrinks,
-  getDrinkBySlug,
-  getDrinksByTag,
-  getAllTags,
-} from "./implementation/queries.server";
-
-// Mutations (orchestrate DB + ImageKit + Fastly + search cache)
-export { createDrink, updateDrink, deleteDrink } from "./implementation/mutations.server";
-
 // Validation
 export { drinkFormSchema } from "./implementation/validation";
-
-// View helpers
-export { withPlaceholderImages } from "./implementation/placeholder-images.server";
-export { parseImageUpload } from "./implementation/parse-image-upload.server";
-export { markdownToHtml } from "./implementation/markdown.server";
 
 // Cache
 export { getSurrogateKeyForTag } from "./implementation/tags";
