+++ app/modules/search/index.server.ts
@@ -1,0 +1,1 @@
+export { searchDrinks, purgeSearchCache } from "./implementation/search.server";
