+++ app/modules/search/index.ts
@@ -1,4 +1,3 @@
-export { searchDrinks, purgeSearchCache } from "./implementation/search.server";
 export { SearchForm } from "./ui/search-form";
 export { NoSearchTerm } from "./ui/no-search-term";
 export { NoDrinksFound } from "./ui/no-drinks-found";
