+++ app/routes/_app.tags.$tag.tsx
@@ -4,12 +4,8 @@ import { cacheHeader } from "pretty-cache-header";
 import { invariantResponse } from "@epic-web/invariant";
 import { defaultPageDescription, defaultPageTitle } from "#/app/core/config";
 import { getLoaderDataForHandle } from "#/app/core/utils";
-import {
-  DrinkList,
-  getDrinksByTag,
-  getSurrogateKeyForTag,
-  withPlaceholderImages,
-} from "#/app/modules/drinks";
+import { getDrinksByTag, withPlaceholderImages } from "#/app/modules/drinks/index.server";
+import { DrinkList, getSurrogateKeyForTag } from "#/app/modules/drinks";
 import { getEnvVars } from "#/app/utils/env.server";
 import type { AppRouteHandle } from "#/app/types";
 import type { Route } from "./+types/_app.tags.$tag";
