+++ app/routes/login.tsx
@@ -1,5 +1,5 @@
 import { redirect } from "react-router";
-import { authenticator, getSession, safeRedirectTo } from "#/app/modules/auth";
+import { authenticator, getSession, safeRedirectTo } from "#/app/modules/auth/index.server";
 import type { Route } from "./+types/login";
 
 export async function loader({ request }: Route.LoaderArgs) {
